requirements.txt
requests
pandas
sqlalchemy
psycopg2-binary
python-dotenv
prometheus_client
uvicorn
fastapi